// packages/lune-admin-ui/src/lib/shop/components/details/form/form-schema.ts
import z from 'zod';

import { FormMessages } from '@/shared/forms/form-messages';

export const ShopDetailsFormSchema = z.object({
  name: z.string().min(1, FormMessages.required),
  email: z.email(FormMessages.invalidEmail),
  phoneNumber: z.string().min(1, 'Phone number is required'),
  logo: z.string().optional(),
  fb: z.string().optional(),
  tw: z.string().optional(),
  ig: z.string().optional()
});